Add files via upload
Development of Rural Education in India Majority of India still lives in villages and so the topic of rural education in India is of utmost importance. The major source of education in rural areas is the Government Schools.But the schools teaching institute under government lacks in vast majority.

In Government schools, most of the students face many issues like poor infrastructure, access and poor quality of educators.

We came up with the web based GUI solution where the students can come up with the complaints they are facing in schools. All the information is collected and analyzed. The analyzed information consists of the performance of teachers and the infrastructure related problems like poor ceiling, washrooms, etc. and any other issues faced in school, This information is sent to related concerned higher authority where necessary steps have to be taken in stipulated time. Also the portal will have access to all the teacher database to monitor on teachers and their activity.